Update retention policy implementation
Do not reverse the order of the snapshots aggregated for one retention
policy as the bahavior isn't what I want.

In the case of a simplified policy of one hour during one day then one
day for a week I would have the following snapshots:

```
2018-05-05-230000
2018-05-05-220000
2018-05-05-......
2018-05-05-010000
2018-05-05-000000
```

Taking the first from the reversed date order would retain the
`2018-05-05-230000` as the snapshot for the day instead of
`2018-05-05-000000` that would allow you to go further back in time.