Do not check subvolume delete return code
It should not be checked as we don't want to impact the rest of the
execution. At first there was a try/except around it to return a sucess
or faillure but it wasn't needed to te try/ecept block was removed but
the check return code flag was still set.